+++ src/main/kotlin/com/vanniktech/maven/publish/MavenPublishPluginExtension.kt
@@ -1,7 +1,5 @@
 package com.vanniktech.maven.publish
 
-import com.vanniktech.maven.publish.nexus.NexusOptions
-import org.gradle.api.Action
 import org.gradle.api.Project
 
 /**
@@ -39,13 +37,4 @@ abstract class MavenPublishPluginExtension(
    * @Since 0.9.0
    */
   var releaseSigningEnabled: Boolean = project.findOptionalProperty("RELEASE_SIGNING_ENABLED")?.toBoolean() ?: true
-
-  /**
-   * Allows to promote repositories without connecting to the nexus instance console.
-   * @since 0.9.0
-   */
-  @Deprecated("Nexus is configured automatically")
-  fun nexus(action: Action<NexusOptions>) {
-    project.baseExtension.nexusOptions(action)
-  }
 }